Derive colours from palette in init
This assumes that whatever stylesheet/theming dependency that is in used has
set an application palette.

If not, then the colours used will be Qt5's default colours.

The thing I noticed with Qt5 apparently deriving a palette from the active
stylesheet as part of the `paintEvent` call turns out to be less than useful,
as the process spits out different colours on different OSes.